+++ common/src/main/java/org/sonarlint/intellij/common/nodejs/NodeJsProvider.kt
@@ -17,7 +17,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonarlint.intellij.common.util;
+package org.sonarlint.intellij.common.nodejs
 
 import com.intellij.openapi.extensions.ExtensionPointName
 import com.intellij.openapi.project.Project
