+++ nodejs/build.gradle.kts
@@ -7,9 +7,11 @@ plugins {
 
 intellij {
     if (!ultimateHome.isNullOrBlank()) {
+        println("Using local installation of Ultimate: $ultimateHome")
         localPath.set(ultimateHome)
         localSourcesPath.set(ultimateHome)
     } else {
+        println("No local installation of Ultimate found, using version $intellijUltimateBuildVersion")
         version.set(intellijUltimateBuildVersion)
     }
     plugins.set(listOf("JavaScript"))
